Music: Added AJAX fetching of reviews. It is so far *not* ideal. (Read commit message body)
Trying to use AJAX for this projects yields several issues:
    * JavaScript does not allow cross origin AJAX unless the host server sets a header for it -- which Daily Bruin does not have
    * A JSONP workaround does not work because JSONP requires a GET request (our current system for grabbing reviews is via POST)
    * The JSON returned by the DB JSON endpoint is corrupted

I managed to get around all of these issues by sending an AJAX response to the Django server, which itself then sends the POST request
to the DB server. The Django server also strips the JSON response of various corrupt / weird parts, and finally returns the content
to the browser.

Unfortunately this approach is *slow*, and I am unsure yet how to speed it up.